\name{create.connection}
\alias{create.connection}
\title{create connection}
\description{Returns a MySQL connection }
\usage{
create.connection(machine.name, machines)
}
\arguments{
  \item{machine.name}{the machine to which the connection is to be made}
  \item{machines}{
}
}



\author{Soumendra Prasad Dhanee (soumendra@gmail.com)}





