\name{csvtodb}
\alias{csvtodb}
\title{csvtodb}
\description{the functions takes all the csv files in the given folder and stores them as different tables in the given database.}
\usage{
csvtodb(folder, user)
}
\arguments{
  \item{folder}{folder which contains the csv files to be moved to database}
  \item{user}{
}
}



\author{Soumendra Prasad Dhanee (soumendra@gmail.com)}





