\name{dbtocsv}
\alias{dbtocsv}
\title{dbtocsv}
\description{the function returns the tables in the given database in the form of csv files, one for each table in the given folder.}
\usage{
dbtocsv(user, folder = getwd())
}
\arguments{
  \item{user}{user name to create connection to the database}
  \item{folder}{folder to which the database is to be copied in the form of csv files}
}



\author{Soumendra Prasad Dhanee (soumendra@gmail.com)}





